requirements.txt
APScheduler==3.6.3
astroid==2.4.2
autopep8==1.5.4
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib==1.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pymongo==3.11.2
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
